data: elm: images: fix double data file definitions leading to races
For a while now I had strange races during make install on Jenkins where some
image files either already existed or chmod was called on files not there yet.

It took some digging but it turns out commit
8dcd5207cc61e604efc942a62137bdeda4060aa0 broke this as a side effect. The black
magic that already installed the files JP refers to is sitting in
data/Makefile.am. It is plain autofoo stuff and just includes the images
Makefile to get the list of files.

JP was correct though that the glayer files have not been installed into the
correct subdir. Fixing this as well here.