elementary: Use software engine for wayland client pointers
There seem to be an issue with the ecore_evas_wayland_egl engine when
using them for cursors. The issue is that a black square shows up
behind the mouse pointer. This does not happen with the wayland_shm
engine so use wayland_shm engine for mouse pointers (for now) until
this can be sorted out.

Signed-off-by: Chris Michael <cp.michael@samsung.com>